+++ EmployeeManagement.Business/Services/TeamService.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using EmployeeManagement.Business.Exceptions;
 using EmployeeManagement.Business.Validation;
 using EmployeeManagement.Common.Dtos.Teams;
 using EmployeeManagement.Common.Interfaces;
@@ -29,10 +30,15 @@ namespace EmployeeManagement.Business.Services
         public async Task<int> CreateTeamAsync(TeamCreate teamCreate)
         {
             await _teamCreateValidator.ValidateAndThrowAsync(teamCreate);
+
             // filter
             Expression<Func<Employee, bool>> employeeFilter = (employee) => teamCreate.Employees.Contains(employee.Id);
-            
             var employees = await _employeeRepository.GetFilteredAsync(new[] {employeeFilter}, null, null);
+            
+            var missinEmployee = teamCreate.Employees.Where(id => !employees.Any(existing => existing.Id == id));
+            if (missinEmployee.Any())
+                throw new EmployeesNotFoundException(missinEmployee.ToArray());
+
             var entity = _mapper.Map<Team>(teamCreate);
             entity.Employees = employees;
             await _teamRepository.InsertAsync(entity);
@@ -43,6 +49,10 @@ namespace EmployeeManagement.Business.Services
         public async Task DeleteTeamAsync(TeamDelete teamDelete)
         {
             var entity = await _teamRepository.GetByIdAsync(teamDelete.Id);
+
+            if (entity == null)
+                throw new TeamNotFoundException(teamDelete.Id);
+
             _teamRepository.Delete(entity);
             await _teamRepository.SaveChangesAsync();
         }
@@ -50,12 +60,20 @@ namespace EmployeeManagement.Business.Services
         public async Task<TeamGet> GetTeamAsync(int id)
         {
             var entity = await _teamRepository.GetByIdAsync(id, (team) => team.Employees);
+
+            if (entity == null)
+                throw new TeamNotFoundException(id);
+
             return _mapper.Map<TeamGet>(entity);
         }
 
         public async Task<List<TeamGet>> GetTeamsAsync()
         {
             var entities = await _teamRepository.GetAsync(null,null, (team) => team.Employees);
+
+            if (entities == null)
+                throw new TeamNotFoundException();
+
             return _mapper.Map<List<TeamGet>>(entities);
         }
 
@@ -66,7 +84,15 @@ namespace EmployeeManagement.Business.Services
             Expression<Func<Employee, bool>> employeeFilter = (employee) => teamUpdate.Employees.Contains(employee.Id);
 
             var employees = await _employeeRepository.GetFilteredAsync(new[] { employeeFilter }, null, null);
+
+            var missinEmployee = teamUpdate.Employees.Where(id => !employees.Any(existing => existing.Id == id));
+            if (missinEmployee.Any())
+                throw new EmployeesNotFoundException(missinEmployee.ToArray());
+
             var existingEntity = await _teamRepository.GetByIdAsync(teamUpdate.Id, (team) => team.Employees);
+            if(existingEntity == null)
+                throw new TeamNotFoundException(teamUpdate.Id);
+
             _mapper.Map(teamUpdate, existingEntity);
             existingEntity.Employees = employees;
             _teamRepository.Update(existingEntity);
